% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{allEffects.lmm}
\alias{allEffects.lmm}
\title{Functions For Constructing Effect Displays}
\usage{
\method{allEffects}{lmm}(mod)
}
\arguments{
\item{mod}{a model fit by lmmaov model object}
}
\value{
a plot
}
\description{
Functions For Constructing Effect Displays
}
