% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFactors.R
\name{allFactors}
\alias{allFactors}
\title{Title}
\usage{
allFactors(data, vars)
}
\arguments{
\item{data}{a \code{\link{data.frame}} containing columns of variables.}

\item{vars}{a \code{\link{vector}} of variable names existing in the dataframe to perform normality tests on.}
}
\value{

}
\description{
Coerce variables to factors
}
\examples{
Heart <- allFactors(Heart, vars=c("drug","time","patient"))
}
