% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{combine.report}
\alias{combine.report}
\title{Combine several fpr objects into one report}
\usage{
combine.report(objects, style = "multiline", plots = FALSE, round = 4, ...)
}
\arguments{
\item{objects}{a list of objects produced by fpr functions to combine into one report.}

\item{style}{a pandoc table style: \code{simple, multiline, grid, or rmarkdown}.}

\item{plots}{plots a TRUE/FALSE boolean value indicating whether or not to perform include plots in the report.}

\item{round}{the number of digits to round too when producing the report.}

\item{...}{additional itemes to be passed to \code{\link{Pandoc}} for processing.}
}
\value{
a printed pander tables and optional figures.
}
\description{
Combine several fpr objects into one report
}
