% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics on variables within a data frame}
\usage{
descriptives(
  data,
  vars,
  groupby = NULL,
  conf.level = 0.95,
  medianCI = FALSE,
  R = 2000,
  round = NULL
)
}
\arguments{
\item{data}{A \code{\link{data.frame}} containing columns of variables.}

\item{vars}{A \code{\link{vector}} of variable names existing in the dataframe to perform descriptive statistics on.}

\item{groupby}{An optional \code{\link{vector}} containing factor names within a dataframe used for by group processing.}

\item{conf.level}{The confidence level for mean and median confidence intervals. The default value is 0.95.}

\item{medianCI}{A TRUE/FALSE boolean value indicating whether or not to produce bootstrapped confidence intervals on the median (computationally intensive).}

\item{R}{The number of bootstrap replications to use if medianCI = TRUE. Otherwise has no effect.}

\item{round}{The number of digits to round the final result.}
}
\value{
A \code{\link{list}} containing descriptive statistics for each variable listed in the \code{vars} statement.
}
\description{
The \code{descriptives} function is used to perform descriptive statistics on a set of variables in a \code{\link{data.frame}}.
This can include descriptive statistics by group when including a vector of categorical variables in the groupby= argument.
}
\examples{
descriptives(data=mtcars, vars=c("mpg","disp"), groupby=c("vs","am"))
}
