% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{Effect.lmmaov}
\alias{Effect.lmmaov}
\title{Title}
\usage{
\method{Effect}{lmmaov}(focal.predictors, mod, ...)
}
\arguments{
\item{focal.predictors}{a lmmaov model}

\item{mod}{model fit by lmmaov}

\item{...}{not used}
}
\value{
a plot
}
\description{
Title
}
