% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{emmeans}
\alias{emmeans}
\title{Estimated marginal means (Least-squares means)}
\usage{
emmeans(object, specs, ...)
}
\arguments{
\item{object}{a lmmaov fitted model object}

\item{specs}{A character vector specifying the names of the predictors over which EMMs are desired. specs may also be a formula or a list (optionally named) of valid specs.}

\item{...}{Additional emmeans specifications}
}
\description{
Estimated marginal means (Least-squares means)
}
