% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalvar.R
\name{equalvar}
\alias{equalvar}
\title{Equality of Variances}
\usage{
equalvar(data, formula, center = "median", plots = FALSE)
}
\arguments{
\item{data}{A \code{\link{data.frame}} containing columns of variables.}

\item{formula}{An object of class \code{\link{formula}}.  The left side of the formula should be a numeric continuous response.  The right side of the formula should consist of categorical factors. The variables on the right-hand-side of the model must all be factors and must be completely crossed.}

\item{center}{The name of a function to compute the center of each group; mean gives the original Levene's test; the default, median, provides a more robust test.}

\item{plots}{A TRUE/FALSE boolean indicating whether or not to produce boxplots}
}
\value{
Returns a \code{\link{data.frame}} meant to be printed showing the results of the test.
}
\description{
Equality of Variances
}
\note{
This function primarily utilizes functionality from the \code{\link{car}} package
}
\examples{
equalvar(data=mtcars, mpg~am*vs, center="median")
}
\references{
Fox, J. (2016) Applied Regression Analysis and Generalized Linear Models, Third Edition. Sage.

Fox, J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.
}
