% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart}
\alias{heart}
\title{Modelling covariance structure in the analysis of repeated measures data}
\format{
An object of class \code{data.frame} with 120 rows and 5 columns.
}
\source{
Littel, R. C., Milliken, G. A., Stroup, W. W., and Wolfinger, R. D. (1996), SAS System for Mixed Models, SAS Institute (Data Set 3.5).
}
\usage{
heart
}
\description{
The HR data frame has 120 rows and 5 columns of the heart rates of patients under one of three possible drug treatments.

\describe{
\item{patient}{Patient ID}
\item{drug}{the drug treatment - a factor with levels a, b and p where p represents the placebo.}
\item{basehr}{the patient's base heart rate}
\item{HR}{the observed heart rate at different times in the experiment}
\item{time}{the time of the observation (in minutes)}
}
}
\keyword{datasets}
