% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmaovTMB.R
\name{lmmaovTMB}
\alias{lmmaovTMB}
\title{Fit Models with TMB}
\usage{
lmmaovTMB(
  data,
  fixed,
  random,
  repeated = NULL,
  units = NULL,
  rcov = "vc",
  type = 3,
  REML = TRUE,
  control = glmmTMBControl(optimizer = optim, optArgs = list(method = "BFGS"))
)
}
\arguments{
\item{data}{a \code{\link{data.frame}} containing columns of variables.}

\item{fixed}{a two-sided linear formula object describing the fixed-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right,}

\item{random}{a one-sided formula describing the random effects}

\item{repeated}{the variable name in the \code{\link{data.frame}} representing the repeated measure}

\item{units}{the variable name in the \code{\link{data.frame}}}

\item{rcov}{the R-side variance/covariance matrix diag (diagonal, heterogeneous variance)
ar1 (autoregressive order-1, homogeneous variance)
cs (compound symmetric, heterogeneous variance)
ou (* Ornstein-Uhlenbeck, homogeneous variance)
exp (* exponential autocorrelation)
gau (* Gaussian autocorrelation)
mat (* Matérn process correlation)
toep (* Toeplitz)
Structures marked with * are experimental/untested. See vignette("covstruct", package = "glmmTMB")< for more information.}

\item{type}{type of test, "II", "III", 2, or 3. Roman numerals are equivalent to the corresponding Arabic numerals.}

\item{REML}{whether to use REML estimation rather than maximum likelihood.}

\item{control}{parameters, see glmmTMBControl.}
}
\value{

}
\description{
Fit a linear mixed model using TMB including r-side covariance structures using containment Anova degrees of freedom tests.
}
\examples{
data(heart)
heart<-allFactors(heart, vars=c("drug","time","patient"))
(m1<-lmmaovTMB(data=heart, fixed=HR~drug*time+basehr, random=~patient:drug, repeated="time", units="patient", rcov="ar1"))
}
