% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report}
\alias{report}
\title{Produce Output using Pandoc tables}
\usage{
report(object, style = "multiline", plots = FALSE, round = 4, ...)
}
\arguments{
\item{object}{a R object from an fpr function.}

\item{style}{a pandoc table style: \code{simple, multiline, grid, or rmarkdown}.}

\item{plots}{a TRUE/FALSE boolean value indicating whether or not to perform include plots in the report.}

\item{round}{The number of digits to round too when producing the report.}

\item{...}{additional itemes to be passed to \code{\link{Pandoc}} for processing.}
}
\value{
a printed pander table and optional figures
}
\description{
The \code{report} function will produce ASCII formated output. This includes both custom and pandoc tables and figures.
}
\examples{
m1<-descriptives(data=mtcars, vars=c("mpg","disp"), groupby=c("vs","am"))
report(m1)
}
